# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para controle de usuários e invoices com autenticação JWT.",
        "title": "Pix Generation API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Konachse"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
...
        "/signature": {
            "get": {
                "description": "Retorna todas as assinaturas cadastradas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signature"
                ],
                "summary": "Busca todas as assinaturas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Signature"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria uma nova assinatura vinculada a um cliente e centro de custo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signature"
                ],
                "summary": "Cria uma nova assinatura",
                "parameters": [
                    {
                        "description": "Dados da assinatura",
                        "name": "signature",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SignatureReceive"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Erro de validação ou cliente/centro de custo não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno ao criar assinatura",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove uma assinatura pelo SignatureID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signature"
                ],
                "summary": "Deleta assinatura",
                "parameters": [
                    {
                        "description": "SignatureID da assinatura",
                        "name": "signature",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SignatureDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/signature/cliente": {
            "post": {
                "description": "Retorna assinaturas com base no ClienteID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signature"
                ],
                "summary": "Busca assinatura pelo ClienteID",
                "parameters": [
                    {
                        "description": "ClienteID do cliente",
                        "name": "cliente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SignatureClienteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Signature"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/signature/id/{id}": {
            "get": {
                "description": "Retorna uma assinatura pelo SignatureID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signature"
                ],
                "summary": "Busca assinatura por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SignatureID da assinatura",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Signature"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de uma assinatura",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signature"
                ],
                "summary": "Atualiza assinatura",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SignatureID da assinatura",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novos dados da assinatura",
                        "name": "signature",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SignatureReceive"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Signature": {
            "type": "object",
            "properties": {
                "centro_custo": {
                    "type": "string"
                },
                "cliente_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "dia_lancamento": {
                    "type": "integer"
                },
                "dia_vencimento": {
                    "type": "integer"
                },
                "emitido_este_mes": {
                    "type": "boolean"
                },
                "qtd_parcelas": {
                    "type": "integer"
                },
                "signature_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "valor_operacao": {
                    "type": "number"
                },
                "vencido_este_mes": {
                    "type": "boolean"
                }
            }
        },
        "model.SignatureClienteRequest": {
            "type": "object",
            "required": [
                "cliente_id"
            ],
            "properties": {
                "cliente_id": {
                    "type": "string"
                }
            }
        },
        "model.SignatureDeleteRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "model.SignatureReceive": {
            "type": "object",
            "required": [
                "centro_custo",
                "cliente_id",
                "dia_lancamento",
                "dia_vencimento",
                "qtd_parcelas",
                "valor_operacao"
            ],
            "properties": {
                "centro_custo": {
                    "type": "string"
                },
                "cliente_id": {
                    "type": "string"
                },
                "dia_lancamento": {
                    "type": "integer"
                },
                "dia_vencimento": {
                    "type": "integer"
                },
                "qtd_parcelas": {
                    "type": "integer"
                },
                "valor_operacao": {
                    "type": "number"
                }
            }
        },
...
    }
}